Changed total depth column to float to allow for some entries to have decimals. Also updated the layer loading function to reducing print out while making it more informative and increasing speed

+++ snowxsql/data.py
@@ -77,7 +77,7 @@ class LayerData(SingleLocationData, Base):
     slope_angle = Column(Integer)
     aspect = Column(Integer)
     air_temp = Column(Float)
-    total_depth = Column(Integer)
+    total_depth = Column(Float)
     surveyors = Column(String(50))
     weather_description = Column(String(50))
     precip = Column(String(50))
